requirements/base.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coverage==6.2
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==4.0
django-allauth==0.43.0
django-autoslug==1.9.8
django-cache-url==3.2.3
django-ckeditor==6.2.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-model-utils==4.2.0
environs==8.0.0
factory-boy==3.2.1
Faker==11.3.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
marshmallow==3.14.1
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
